requirements.txt
python-dotenv==0.18.0
Flask==2.0.2
Werkzeug==2.0.3